+++ xcube/core/store/fs/store.py
@@ -11,9 +11,6 @@ from threading import RLock
 from typing import (
     Optional,
     Any,
-    Tuple,
-    List,
-    Dict,
     Union,
     Callable,
 )